Upload bug fix

+++ Sponge/ViewModel/Upload.cs
@@ -9,7 +9,7 @@
     }
     public class FileModel
     {
-        public decimal TemplateID { get; set; }
+        public int TemplateID { get; set; }
         public string FileName { get; set; }
         public DateTime? PERIOD_FROM { get; set; }
         public DateTime? PERIOD_TO { get; set; }
@@ -19,16 +19,16 @@
         public string DocumentID { get; set; }
         public string DocumentFileNAME { get; set; }
         public string ErrorMessage { get; set; }
-        public decimal? DOCTEMPLATEID { get; set; }
+        public int? DOCTEMPLATEID { get; set; }
         public List<FileModel> ListTemplateDetails;
-        public decimal RoleID { get; set; }
-        public decimal ConFigId { get; set; }
+        public int RoleID { get; set; }
+        public int ConFigId { get; set; }
         public string UploderUserId { get; set; }
         public string UploderEmailId { get; set; }
         public string UserName { get; set; }
         public string OnTime { get; set; }
         public string ForTime { get; set; }
-        public decimal? SubFunctionID { get; set; }
+        public int? SubFunctionID { get; set; }
 
         public string ApproverID { get; set; }
         public String ApproverName { get; set; }
@@ -37,7 +37,7 @@
     public class TemplateFile
     {
         public string FileName { get; set; }
-        public decimal TemplateID { get; set; }
+        public int TemplateID { get; set; }
         public string ErrorMessage { get; set; }
         public string ErrorType { get; set; }
         public List<TemplateFile> TeplateFileList;
